Refs #1 unit test
